fix: Change primitive type of totalWinningMoney
- Fix produceResultForEachTicket method logic

+++ src/main/java/lotto/domain/Rank.java
@@ -13,9 +13,9 @@ public enum Rank {
     MISS(0, 0);
 
     private final int countOfMatch;
-    private final int winningMoney;
+    private final long winningMoney;
 
-    Rank(int countOfMatch, int winningMoney) {
+    Rank(int countOfMatch, long winningMoney) {
         this.countOfMatch = countOfMatch;
         this.winningMoney = winningMoney;
     }
@@ -58,7 +58,7 @@ public enum Rank {
         return countOfMatch;
     }
 
-    public int getWinningMoney() {
+    public long getWinningMoney() {
         return winningMoney;
     }
 }
